requirements.txt
fastapi~=0.111.0
httpx~=0.27.0
beautifulsoup4~=4.12.3
pydantic~=2.8.2
redis~=5.0.7
aiofiles
python-dotenv~=1.0.1
SQLAlchemy~=2.0.31
requests~=2.32.3